Skip installation if resolved version is a default gem
Running `gem install <name>:<version>` where `<name>-<version>` is a
default gem has the potential to break things because of little
divergences on how ruby-core installs default gems vs how rubygems
installs default gems.

For example, the `bundler` gem installed by ruby-core has its
executables installed on a "libexec" folder inside the gem installation
folder (it's the only thing included there because the other files
belong in the stdlib core folder). However, when installed as a regular
gem this folder is called "exe". If we allow overwriting the default gem
with a regular installation, the "libexec" folder will be renamed to
"exe", but the default gemspec will still point to the original
"libexec" folder. This can cause issues when changing `GEM_PATH` and
`GEM_HOME` (ignoring the regular gem installed), because the "libexec"
folder will no longer be found.

The easiest way to prevent this mess seems to skip installation of gems
if the same resolved version is already installed as a default gem.