Remove no longer needed `Psych::PrivateType` cleanup
This old bug used to affect the `rubyforge_project` field in serialized
gemspecs. However, this field has been removed and it's no longer
present in marshal loaded gemspecs.

So, while the constant is still present in these marshalled gemspecs, so
we still need to make sure it still exists, we no longer need to clean
it up.